Bug Fixes
Corrected templeting errors on the home page which were causing odd effects with the AuthService integration
Improved some syntax issues
Simplified home page service interaction code

Note on last point:
the code that was removed from home.ts was previously deemed essential due to differences with Ionic pages compared to Angular Components. This was incorrect. In actuality, when auth.user$ or auth.googleSignIn was being called on the HTML it was making a ref to the ng-template #auth and not the auth ref created in the TS which specifies the AuthService. Most of these changes are simply correcting that issue